Three signed countdown bytes: volleyShotsRemaining, hitCountdown, laserLevel
Each was typed u8 and then read back exclusively through a cast, and in the
squadron's case written through *(s8*)& as well:

  678 ArwSquadronState.volleyShotsRemaining -- seeded from shotsPerVolley,
      decremented per shot, drained on (s8)x <= 0
  592 KtRexState.hitCountdown -- seeded to 3, decremented per hit, tested
      (s8)x != 0 / <= 0 / < 1
  666 ArwingState.laserLevel -- tested (s8)x == 2 / == 1 / < 2 and
      incremented

Typing all three s8 removes ten casts and one address pun with all three
objects byte-identical, so retail sign-extends every one of these reads.
This is the opposite of ArwingState.health/maxHealth in the same file, whose
reads retail zero-extends (arwarwing_addHealth emits lbz; add; extsb; stb) --
those casts are load-bearing and stay.

+++ src/dlls/objects/592_KT_Rex/KT_Rex.c
@@ -1046,7 +1046,7 @@ void ktrex_updateContactEffects(GameObject* obj, KTRexRuntime* runtime)
         return;
     }
     contactPoints = *(f32**)((u8*)((ObjAnimComponent*)obj)->banks[((ObjAnimComponent*)obj)->bankIndex] + 0x50);
-    if ((s8)runtime->hitCountdown != 0 && (hitType == 3 || hitType == 2) &&
+    if (runtime->hitCountdown != 0 && (hitType == 3 || hitType == 2) &&
         (gKTRexState->timerFA & 0x10) != 0 && hit == 5)
     {
         gKTRexEffectSpawnWork.posX = playerMapOffsetX + (pt = contactPoints + hitType * 4)[1];
@@ -1064,7 +1064,7 @@ void ktrex_updateContactEffects(GameObject* obj, KTRexRuntime* runtime)
         {
             runtime->hitCountdown = 0;
         }
-        if ((s8)runtime->hitCountdown <= 0)
+        if (runtime->hitCountdown <= 0)
         {
             runtime->hitCountdown = 0;
             gKTRexState->timerFA &= ~0x10;
@@ -1094,7 +1094,7 @@ void ktrex_updateContactEffects(GameObject* obj, KTRexRuntime* runtime)
                     (StaffCollisionColorArgs*)msg.w);
         gKTRexContactEffectCooldown = 0x3c;
     }
-    if ((s8)runtime->hitCountdown < 1)
+    if (runtime->hitCountdown < 1)
     {
         runtime->hitCountdown = 0;
     }
